client order form & env setup

+++ src/main/java/com/bondarmih/ticketbooking/data/entity/User.java
@@ -17,7 +17,7 @@ public class User {
     @Id
     @GeneratedValue(strategy=GenerationType.SEQUENCE)
     @Column(name = "id")
-    private Integer id;
+    private long id;
 
     @Column(name = "name", unique = true)
     private String name;
@@ -35,11 +35,11 @@ public class User {
     @OneToMany(mappedBy = "user")
     private Set<Order> orders;
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
